Nowe rzeczy

+++ Network/IEquipableStats.cs
@@ -8,7 +8,7 @@ namespace NetworkProject
     public interface IEquipableStats : IStats
     {
         int Lvl { get;}
-        float AttackSpeed { get; set; }
+        int AttackSpeed { get; set; }
         int MinDmg { get; set; }
         int MaxDmg { get; set; }
         int Defense { get; set; }
